reorder and use IssueDescription component
