+++ src/components/CheckDrillDown/CheckDrillDown.test.tsx
@@ -4,7 +4,7 @@ import userEvent from '@testing-library/user-event';
 import CheckDrillDown, { CheckDrillDownProps } from './CheckDrillDown';
 import { CheckSummaries, Severity } from 'types';
 import { getEmptyCheckSummary, getEmptyCheckTypes } from 'api/api';
-import { renderWithRouter } from './test/utils';
+import { renderWithRouter } from '../test/utils';
 
 describe('Components/CheckDrillDown', () => {
   let checkSummaries: CheckSummaries;
